+++ src/main/java/lotto/LottoApplication.java
@@ -1,9 +1,9 @@
 package lotto;
 
 import lotto.domain.Lotto;
-import lotto.domain.LottoFactory;
 import lotto.domain.LottoStatisticsResult;
 import lotto.domain.Money;
+import lotto.utils.LottoFactory;
 import lotto.utils.LottoStatisticsCalculator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
