+++ src/test/java/lotto/utils/LottoFactoryTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
